Separate into nested classes logic in the `FailureHelper` class.
